+++ .gitignore
@@ -1,7 +1,9 @@
 ### IntelliJ IDEA ###
 out/
+.idea/
 !**/src/main/**/out/
 !**/src/test/**/out/
+*.iml
 
 ### Eclipse ###
 .apt_generated
