+++ src/Choice.java
@@ -1,16 +1,21 @@
+import java.io.File;
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
 
 public class Choice {
-    List<Object> contents = new ArrayList<>();
+    String text;
+    File scene;
+    Choice(HashMap<String, String> choice) {
+        text = choice.get("text");
+        scene = new File(choice.get("scene"));
+    }
 
-    // Might change to a list of effects
-    List<Object> effect0 = new ArrayList<>();
-    List<Object> effect1 = new ArrayList<>();
+    public String getText() {
+        return text;
+    }
 
-    Choice(List<Object> contents, List<Object> effect0, List<Object> effect1) {
-        this.contents = contents;
-        this.effect0 = effect0;
-        this.effect1 = effect1;
+    public File getScene() {
+        return scene;
     }
 }
