+++ src/Scene.java
@@ -2,23 +2,24 @@ import java.io.*;
 import java.util.*;
 
 public class Scene {
-    private Metadata metadata;
+    private File background;
     private List<Dialog> dialogs = new ArrayList<>();
     private int currentIndex = 0;
 
     Scene (File sceneFile) {
         Parser parser = new Parser(sceneFile);
-        metadata = parser.extractMetadata();
+        background = parser.extractBackground();
         dialogs = parser.extractDialog();
+    }
+
+    public File getBackground() {
+        return background;
     }
 
     public boolean isOver() {
         return currentIndex == dialogs.size();
     }
 
-    public Metadata getMetadata() {
-        return metadata;
-    }
     public Dialog getCurrentDialog() {
         return dialogs.get(currentIndex);
     }
